requirements.txt
torch>=2.1.0
transformers>=4.36.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0

